move new character form to another file
